+++ app/assets/stylesheets/custom.scss
@@ -42,7 +42,7 @@ $gray-medium-light: #eaeaea;
   }
 }
 
-.subjects {
+.subject_table {
   width: auto;
   th {
     text-align: center;
@@ -50,12 +50,11 @@ $gray-medium-light: #eaeaea;
   td {
   }
   .subject_name {
-    max-width: 200px;
-    min-width: 100px;
+    width: 200px;
   }
   .score {
     text-align: center;
-    max-width: 60px;
+    width: 60px;
   }
 }
 
@@ -90,4 +89,11 @@ td.my-button {
 tr.shown td.my-button {
     background: image-url('close.png') no-repeat center;
     background-size: 15px 15px;
+}
+
+
+
+
+select{
+    font-family:"Courier New", Courier, monospace
 }
