+++ app/controllers/static_pages_controller.rb
@@ -1,4 +1,10 @@
+require 'pry'
 class StaticPagesController < ApplicationController
   def usage
+  end
+  
+  def stocked
+    ids = cookies["subjects"].split(',').reject(&:blank?)
+    @subjects = SummarizedSubject.where(id: ids)
   end
 end
