+++ app/views/faculties/index.html.erb
@@ -7,10 +7,7 @@
 <li>使い方がわからない方は<%= link_to "こちら", usage_path  %></li>
 <li>問い合わせは<%= link_to "Twitter", "https://twitter.com/SyllabusPlus", target: :_blank %>より</li>
 </ul>
-
-<div class="cookie-debug">
-</div>
-
+             
 <div class="checkbox-control-area">
   <button type="button" class="checkbox-control-text btn btn-default" onclick="control_checkbox_visible();">学部一覧を閉じる</button>
   <div class="checkbox-control" status="open" style="display: block">
