+++ app/views/layouts/application.html.erb
@@ -59,8 +59,8 @@
         <a class="navbar-brand" href="<%= root_path %>">SyllabusPlus</a>
         <div class="navbar-collapse collapse navbar-responsive-collapse">
           <ul class="nav navbar-nav">
-            <li><%= link_to "使い方", "/usage"  %></li>
-            <!-- <li><%= link_to "Link2", "/path2"  %></li> -->
+            <li><%= link_to "使い方", usage_path  %></li>
+            <li><%= link_to "保存した科目を見る", stocked_path %></li>
             <!-- <li><%= link_to "Link3", "/path3"  %></li> -->
           </ul>
         </div>
