+++ app/views/static_pages/stocked.html.erb
@@ -1,0 +1,13 @@
+<table class="subject_table">
+  <tr>
+    <th class="subject_name">科目名</th>
+    <th>科目コード</th>
+    <th class="score">A</th>
+    <th class="score">B</th>
+    <th class="score">C</th>
+    <th class="score">D</th>
+    <th class="score">F</th>
+    <th class="score">平均評点</th>
+  </tr>
+  <%= render @subjects %>
+</table>
