--- app/views/subjects/_subject.html.erb
@@ -1,19 +1,0 @@
-<tr class="table_data">
-  <td class="subject_name"><%= subject.name %></td>
-  <td><%= link_to subject.code, subject.url %></td>
-  <td class="score"><%= subject.A.round(1) %></td>
-  <td class="score"><%= subject.B.round(1) %></td>
-  <td class="score"><%= subject.C.round(1) %></td>
-  <td class="score"><%= subject.D.round(1) %></td>
-  <td class="score"><%= subject.F.round(1) %></td>
-  <td class="score"><%= subject.mean_score.round(1) %></td>
-  <td style="display:none;"><%= subject.weighted_score %></td>
-</tr>
-<tr class="table_meta">
-  <td colspan="8">
-    <a href="javascript:void(0)" id="text_<%= subject.id %>" onclick="show('<%= subject.id %>');">年度のごとの分布を見る</a>
-    <div class="year_data" id="table_<%= subject.id %>" style="display: none;" status="close">
-      <%= render partial: 'year_data/data_table', locals: { subject: subject } %>
-    </div>
-  </td>
-</tr>
