+++ app/views/summarized_subjects/_summarized_subject.html.erb
@@ -1,0 +1,10 @@
+<tr class="table_data">
+  <td><%= link_to summarized_subject.name, summarized_subject.url %></td>
+  <td><%= summarized_subject.code %></td>
+  <td class="score"><%= summarized_subject.A.round(1) %></td>
+  <td class="score"><%= summarized_subject.B.round(1) %></td>
+  <td class="score"><%= summarized_subject.C.round(1) %></td>
+  <td class="score"><%= summarized_subject.D.round(1) %></td>
+  <td class="score"><%= summarized_subject.F.round(1) %></td>
+  <td class="score"><%= summarized_subject.mean_score.round(1) %></td>
+</tr>
